Refactor LogOnlySignatureTrustEngine to move it to its own class
